Set up a file structure creating the necessary folders for the app and adding initial files in them
Each of the folders in the 'features' folder  contain all the components, all the custom hooks and really all the JavaScript files that are necessary to make each of the feature work.
The 'ui' folder is for components that are more reusable or that don't really belong in any of these features (cart, menu, order, user).
The 'services' folder is basically for reusable code for interacting with an API.
The 'utils' folder is for basically just some helper functions that we can also reuse in multiple places of the application.