Create the AppLayout component, which we will want to use as the parent route of every single other route that we have in our application
And so that's why we placed this app layout element at the top in App.jsx and then we made all the other routes child routes of the app layout. So they are all nested routes now.
And so then inside the parent route, we can use the outlet component to render whatever is the current nested route.
...so we have just one layout all the time. It will never change throughout the app. This is basically our parent route of all other routes.
As we mentioned earlier, since AppLayout doesn't have a path, it is technically called in React Router a layout route now.