Do so when the user enters his name in the home page, to be redirected to the menu with pizzas
Because when the user comes back to the home page after entering his log name, it is the same (as he never logged in), we're changing this so that the home page to be different if the user has already entered its name.
Also, the username of the user will be taken from the redux store from now on.