Implement the delete button. We do it by extracting it in separate component because we want to reuse it in the main menu, too, so that the user could delete the added item immediately if he was added it by mistake, and not only from the cart
For that reason the delete button is rendered on the main menu page when the user adds a pizza. Do so only the delete button to be visible if the current pizza is added to the cart (without the 'add to cart' button).

In addition - fix an unnoticed problem with key item in list.