Setup the updating of cart quantities and add styling for round button
Now we could decrease and increase the value of each pizza.

Do so the item quantity to be rendered between the - and + buttons (this is done also in the main page).

In addition - do so the item be removed if it's quantity is 0 (because before it was getting negative values which mustn't be allowed).

Fix a small bug with key prop.