Create a Thunk middleware by using the built-in CreateAsyncThunk function which is a way of creating a Thunk
The idea here is to implement the feature where our users can use geolocation in order to get their GPS position and their address.

We now have 'Get position' button in the order and when the user clicks on it, if the he allow, its address is found and could be used.