Setup the styling and the position of the 'Get position' button which take the user's coordinates
Do so an error message to be shown on the screen if the user has disallowed its location to be used (we make it through the i button in the navbar, on the left side of the URL of the page (localhost:5173/order/new).

In the form action, when we submit the new order, we will also want to submit it with the actual position data, so with the user's GPS location, because that's going to be really helpful for the company to deliver the pizza. We use the same trick that we used for the card (with the 'hidden' type) to achieve that goal.

Now when the user submit the form, it will be submitted with the GPS location of the user.